Refactoring and cleaning of code
Removed unused code and refactored some classes
